% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomizeText.R
\name{atomizeText}
\alias{atomizeText}
\title{atomizador de texto}
\usage{
atomizeText(text)
}
\arguments{
\item{text}{um texto ou frases em portugues}
}
\value{
um vetor com todas as palavras do texto.
}
\description{
Uma função que realiza a quebra de partes de um texto em portugues em palavras.
}
\examples{

frase <- c(”Alegações extraordinárias exigem evidências extraordinárias.”)

atomizeText(frase)

}
\author{
Luan Cenci
}
