% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractWords.R
\name{extractWords}
\alias{extractWords}
\title{extrator de Termos}
\usage{
extractWords(text, targetWords)
}
\arguments{
\item{text}{um texto ou frases em portugues}

\item{targetWords}{vetores contendo termos ou termos compostos}
}
\value{
retorna todos os termos pedidos nas targetsWords, mesmo se repetirem no texto.
}
\description{
um função que retorna palavras ou termos compostos dentro de um texto em portuges.
}
\examples{

frase <- c(”Alegações extraordinárias exigem evidências extraordinárias.”)

palavra <- "evidência"

atomizeText(frase,palavra)

}
\author{
Luan Cenci
}
