% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{stopWopdsRemoction}
\alias{stopWopdsRemoction}
\alias{stopWordsRemoction}
\title{Eliminador de stopwords}
\usage{
stopWordsRemoction(wordList, stopList)
}
\arguments{
\item{wordList}{listas de palavras que devem ser eliminadas as stopwords}

\item{stopList}{data frame contendo as stopwords em portugues}
}
\value{
um vetor sem as stopWords.
}
\description{
retirar as stopwords em portugues que contem dentro de um texto
}
\examples{

lista <- ("e","o","índio")

stopWordsRemoction(lista,StopList)

}
\author{
Luan Cenci
}
